# openapi/wallet-api.yaml
openapi: 3.1.0
info:
  title: Universal AI Wallet — Wallet API
  version: 1.0.0
  description: Account, balance, keys, and top-ups.
servers:
  - url: https://api.example.com/wallet/v1
paths:
...
  /wallet:
    get:
      operationId: getWallet
      summary: Wallet balance
      tags: [Wallet]
      security:
        - JwtAuth: []
      responses:
        '200':
          description: Balance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Wallet'
  /wallet/transactions:
    get:
      operationId: listTransactions
      summary: Ledger history
      tags: [Wallet]
      security:
        - JwtAuth: []
      parameters:
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: Transaction page
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/LedgerEntry'
                  next_cursor:
                    type: string
...
  /topups/checkout:
    post:
      operationId: createTopupCheckout
      summary: Start Stripe checkout
      tags: [Topups]
      security:
        - JwtAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount_microdollars]
              properties:
                amount_microdollars:
                  type: integer
                  minimum: 500000
      responses:
        '200':
          description: Checkout URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  checkout_url:
                    type: string
                    format: uri
                  payment_intent_id:
                    type: string
                    format: uuid
...
components:
  securitySchemes:
    JwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    OAuthBearer:
      type: http
      scheme: bearer
      description: OAuth2 app access token (JWT with app_install_id claim)
    PartnerAdmin:
      type: apiKey
      in: header
      name: X-Partner-Admin-Token
  schemas:
...
    Wallet:
      type: object
      properties:
        wallet_id:
          type: string
          format: uuid
        balance_microdollars:
          type: integer
        held_microdollars:
          type: integer
        available_microdollars:
          type: integer
        currency:
          type: string
          enum: [USD]
        low_balance_threshold_microdollars:
          type: integer
    LedgerEntry:
      type: object
      properties:
        id:
          type: string
          format: uuid
        entry_type:
          type: string
        amount_microdollars:
          type: integer
        balance_after_microdollars:
          type: integer
        created_at:
          type: string
          format: date-time
...